feat: function snippets
`tsserver` does not create snippets for functions in the completions
list; only upon completion resolve (presumably this is for performance
reasons).

`typescript-language-server` deals with this by manually
constructing the snippet and tweaking the `inputTextFormat`. We copy
that here - as long as `includeCompletionsWithSnippetText = true`.

INFO:
refactor/fix: move default configs to `config.lua` so they can be
referenced outside of just `text_document/did_open`