fix: cache ftplugin on first load
This eliminates repeated reads of ftplugin implementation file on
repeated filetype events. The only reliably persistent store available
to ftplugin scripts that I can think of is the global table, which is
why I'm using it to store loaded ftplugin "object" in. The prefix is a
string gneerated by rolling a fair dice in order to make name collision
highly improbable.

There might be a more efficient way to do this. Please tear this commit
apart if you know of it.