fix: deprecated use of vim.validate
The table argument for `vim.validate` got deprecated in nvim 0.11.

This commit re-uses the existing table argument that used to be passed
to `vim.validate` and re-implements looping over that argument and calls
`vim.validate` as individual function calls for each table value.

This is a minimal change to get rid of the deprecation warning. It does
not change the structure of the code significantly.