feat: better hover
Default `maximumHoverLength` value is `500`. That means hover
information is truncated for complicated types (usually some parts are
omitted and shown as `...`). This is a little bit annoying as it means
you must go deeper into code and analyze type itself while all info
could be in hover. This change solves this problem.

I have intentionally made this non configurable as I think this should
be default in Neovim.